it works on example
it might not work overall tho

+++ lib/cmd.h
@@ -64,7 +64,7 @@ struct Tag {
         ELEM_T fst = POP;                 \
         if (condition) {                  \
                                           \
-            ip = (int) *cmdArg;      \
+            ip = (size_t) *cmdArg;           \
         }                                 \
     })
